Move common code to base sub-module (#59)
