+++ qc_opendrive/checks/__init__.py
@@ -1,4 +1,3 @@
-from . import utils as utils
 from . import semantic as semantic
 from . import geometry as geometry
-from . import models as models
+from . import performance as performance
