+++ qc_opendrive/checks/geometry/road_geometry_parampoly3_normalized_range.py
@@ -6,7 +6,7 @@ from scipy.integrate import quad
 from qc_baselib import IssueSeverity
 
 from qc_opendrive import constants
-from qc_opendrive.checks import utils, models
+from qc_opendrive.base import models, utils
 from qc_opendrive.checks.geometry import geometry_constants
 
 RULE_INITIAL_SUPPORTED_SCHEMA_VERSION = "1.7.0"
