+++ qc_opendrive/checks/semantic/road_lane_access_no_mix_of_deny_or_allow.py
@@ -8,7 +8,7 @@ from lxml import etree
 from qc_baselib import Configuration, Result, IssueSeverity
 
 from qc_opendrive import constants
-from qc_opendrive.checks import utils, models
+from qc_opendrive.base import models, utils
 
 from qc_opendrive.checks.semantic import semantic_constants
